requirements.txt
# Fashion Studio ETL Pipeline - Dependencies
# Python version: 3.9+

# Web scraping
requests~=2.32
beautifulsoup4~=4.12

# Data processing
pandas~=2.2

# Google Sheets API
google-auth~=2.36
google-api-python-client~=2.152

# PostgreSQL database
sqlalchemy~=2.0
psycopg2-binary~=2.9

# Task scheduling (optional)
python-crontab~=3.2

# Testing
pytest~=8.0
pytest-cov~=6.0
